Make the statistics collection safer/possible on macOS.
The stats.py module pulls in a package which might not be installed,
and isn't present on my macOS system - the pyamdgpuinfo package. This
change makes the import silent, and allows it to be auto-detected
as absent, rather than failing to run.

Similarly, the psutil package doesn't appear to provide the
sensors_temperature interface, so this has been made safe if it is not
present.